+++ core/src/options/from_type_param.rs
@@ -14,7 +14,7 @@ pub struct FromTypeParamOptions {
 impl FromTypeParamOptions {
     pub fn new(di: &syn::DeriveInput) -> Result<Self> {
         (FromTypeParamOptions {
-            base: OuterFrom::start(di),
+            base: OuterFrom::start(di)?,
             bounds: None,
             default: None,
         }).parse_attributes(&di.attrs)?